+++ src/com/gourav/leetcode/may30days/CheckIfItIsAStraightLine.java
@@ -35,7 +35,7 @@
 // 
 // Related Topics Array Math Geometry
 	
-package com.gourav.leetcode.may30days.leetcode.editor.en;
+package com.gourav.leetcode.may30days;
 
 //leetcode submit region begin(Prohibit modification and deletion)
 class Solution {
