+++ src/com/gourav/leetcode/may30days/CousinsInBinaryTree.java
@@ -56,7 +56,7 @@
 // 
 // Related Topics Tree Breadth-first Search
 
-package com.gourav.leetcode.may30days.leetcode.editor.en;
+package com.gourav.leetcode.may30days;
 
 //leetcode submit region begin(Prohibit modification and deletion)
