+++ src/com/gourav/leetcode/may30days/FindTheTownJudge.java
@@ -73,7 +73,7 @@
 // 
 // Related Topics Graph
 	
-package com.gourav.leetcode.may30days.leetcode.editor.en;
+package com.gourav.leetcode.may30days;
 
 //leetcode submit region begin(Prohibit modification and deletion)
 class Solution {
