+++ src/com/gourav/leetcode/may30days/JewelsAndStones.java
@@ -29,7 +29,7 @@
 // 
 // Related Topics Hash Table
 	
-package com.gourav.leetcode.may30days.leetcode.editor.en;
+package com.gourav.leetcode.may30days;
 
 
 import java.util.HashMap;
