+++ src/com/gourav/leetcode/may30days/RemoveKDigits.java
@@ -38,7 +38,7 @@
 // 
 // Related Topics Stack Greedy
 	
-package com.gourav.leetcode.may30days.leetcode.editor.en;
+package com.gourav.leetcode.may30days;
 
 import java.util.HashMap;
 import java.util.Stack;
