+++ src/com/gourav/leetcode/may30days/SingleElementInASortedArray.java
@@ -24,7 +24,7 @@
 // Note: Your solution should run in O(log n) time and O(1) space. 
 //
 	
-package com.gourav.leetcode.may30days.leetcode.editor.en;
+package com.gourav.leetcode.may30days;
 
 //leetcode submit region begin(Prohibit modification and deletion)
 class Solution {
