+++ src/com/gourav/leetcode/may30days/ValidPerfectSquare.java
@@ -22,7 +22,7 @@
 // 
 // Related Topics Math Binary Search
 	
-package com.gourav.leetcode.may30days.leetcode.editor.en;
+package com.gourav.leetcode.may30days;
 
 //leetcode submit region begin(Prohibit modification and deletion)
 class Solution {
